aspire: Support inputs and improve container.v0 support
This change adds supports for `inputs` which is a new (optional)
property on resources in the Aspire manifest. These inputs represent
values that deployment tools like `azd` need to provide "somehow".

For our initial support, we support string inputs which contain
metadata on how to generate a default value, based on a minumum random
length. Aspire uses this today to model passwords for cases like
`AddMySqlContainer()` where a deployment tool needs to provide a
password for the root user.

Consider the following AppHost manifest:

```json
{
  "resources": {
    "demoapp": {
      "type": "project.v0",
      "path": "../AspireContainers.DemoApp/AspireContainers.DemoApp.csproj",
      "env": {
        "ConnectionStrings__MySql": "{mysqlabstract.connectionString}"
      },
      "bindings": {
        "http": {
          "scheme": "http",
          "protocol": "tcp",
          "transport": "http",
        }
      }
    },
    "mysqlabstract": {
      "type": "container.v0",
      "image": "mysql:latest",
      "env": {
        "MYSQL_ROOT_PASSWORD": "{mysqlabstract.inputs.password}"
      },
      "bindings": {
        "tcp": {
          "scheme": "tcp",
          "protocol": "tcp",
          "transport": "tcp",
          "containerPort": 3306
        }
      },
      "connectionString": "Server={mysqlabstract.bindings.tcp.host};Port={mysqlabstract.bindings.tcp.port};User ID=root;Password={mysqlabstract.inputs.password}",
      "inputs": {
        "password": {
          "type": "string",
          "secret": true,
          "default": {
            "generate": {
              "minLength": 10
            }
          }
        }
      }
    }
  }
}
```

In this example, we spin up a MySql Container (with a randomly
generated password at least 10 characters long). The container
resource now has an expression with binding references in for its
`connectionString` property, and other resources (like `demoapp`) can
refer to this property to create connections to it.

To support this, we need to teach `azd` how to handle these inputs and
where to store them as well as decide how to the flow into both the
infrastructure and service deployments.

The strategy here for now is to model a new `Inputs` object, which
holds all the values for the given inputs. The object contains a top
level key for every resource with an input and the value is an object
that maps inputs from that resource to their values, eg:

```json
{
  "mysqlabstract": {
     "password": "<value-from-deployment-tool>"
  }
}
```

These are stored in the environment config (so they can vary per
environment) under the "inputs" section.

Values which are marked as `secure` today are not encrypted at rest,
but this is something we will improve before GA, by storing these
values in KeyVault and then reference from there.

We do, however, mark the parameter that gets the value in the bicep
deployment as an `@secure()` parameter, so the values like passwords
are not retained in the deployment log.

As part of this change, some small improvements where made to the
binding expression evaulatior, to allow using `.host` in references to
container apps (we know the value will be the same as the container
app name, due to how networking works in ACA).