recording: Use fixed port in `TestBlobClientGetProperties`
When developing the test I unintentionally set
`recorder.ModeRecordOnly` causing us to recrate the recording per
run. Turning that off exposed an issue where since the `httptest`
package picks a different port each run it meant that the recording
would not match because the port was different, causing us to fail to
playback the test.

For now, use a fixed port for our test server, which ensures we have a
stable `host` value across runs, at the expense of disallowing
multiple copies of the test from running at the same time.