WIP: recording: Add `WithHostMapping` option
This outlines how we might add a WithHostMapping option and use it in
our recorder to allow recording `httptest` based tests in a nice way.

However, it doesn't actually work - while the recordeding file is
correctly updated, this is not yet taken into account when matching,
so forcing playback with the updated recording does not work as
expected.